Factor out to common.rs
